Got rid of BasicVector / La4j [DBchange]
- reason: double is too much
- switched to short
- implemented vector related operations (manhattan norm / manhattan diff) in my class
- put conversion from short[] to byte[] into own class (ArrayUtils)

- rebuild fingerprint database!

+++ baksmali/src/main/java/org/androidlibid/proto/ast/ASTToFingerprintTransformer.java
@@ -45,8 +45,8 @@ public class ASTToFingerprintTransformer {
                 if(lvl2.getType().equals(NodeType.PAR)) params++;
             }
             
-            int llPairs = fromNChoose2(locals); 
-            int ppPairs = fromNChoose2(params); 
+            short llPairs = (short) fromNChoose2(locals); 
+            short ppPairs = (short) fromNChoose2(params); 
             
             print.incrementFeatureBy(llPairs, NodeType.LOC, NodeType.LOC);
             print.incrementFeatureBy(ppPairs, NodeType.PAR, NodeType.PAR);
